Add prodigy to mamenl build. Fix clang5 warnings. (nw)

+++ src/lib/netlist/nl_base.h
@@ -29,8 +29,7 @@
 // ----------------------------------------------------------------------------------------
 
 /*! netlist_sig_t is the type used for logic signals. */
-//using netlist_sig_t = std::uint32_t;
-using netlist_sig_t = std::uint8_t;
+using netlist_sig_t = std::uint32_t;
 
 //============================================================
 //  MACROS / New Syntax
